backend/requirements.txt
fastapi[standard]>=0.112
uvicorn>=0.30
pydantic>=2.9
python-dotenv>=1.0
google-genai>=1.0
